pyproject.toml
[project]
name = "microtutor"
version = "0.1.0"
description = "AI-powered micro-learning personal tutor"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
]
